explainerdashboard/requirements.txt
aiobotocore==2.11.1
aiohttp==3.9.2
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alabaster==0.7.16
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
appdirs==1.4.4
apprise==1.7.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
botocore==1.34.27
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.8.2
colour==0.1.5
contourpy==1.2.0
coolname==2.2.0
croniter==1.4.1
cryptography==42.0.1
cycler==0.12.1
Cython==3.0.8
daiquiri==3.2.5.1
dash==2.15.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.2.0
databricks-cli==0.18.0
dataclasses==0.6
dateparser==1.2.0
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docker==6.1.3
docutils==0.17.1
dtreeviz==2.2.2
entrypoints==0.4
exceptiongroup==1.2.0
explainerdashboard==0.4.5
fastapi==0.109.2
fastapi-restful==0.5.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==2.0.3
flask-simplelogin==0.1.2
Flask-WTF==1.2.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.3
griffe==0.39.1
gTTS==2.5.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
hyperframe==6.0.1
idna==3.6
imagesize==1.4.1
imbalanced-learn==0.12.0
imblearn==0.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==7.34.0
ipywidgets==8.1.2
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.17
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kaggle==1.6.6
kaleido==0.2.1
kiwisolver==1.4.5
kubernetes==27.2.0
lightgbm==3.3.5
llvmlite==0.37.0
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.1
mdurl==0.1.2
minio==7.2.3
mistune==3.0.2
mlflow==2.10.2
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nbsphinx==0.8.12
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.4
numba==0.54.1
oauthlib==3.2.2
optuna==3.5.0
orjson==3.9.12
overrides==7.7.0
oyaml==1.0
packaging==23.2
pandas==1.4.4
pandocfilters==1.5.1
pathspec==0.12.1
pendulum==3.0.0
phik==0.12.4
pillow==10.2.0
plotly==5.19.0
pluggy==1.4.0
ppca==0.0.4
prettytable==3.9.0
prince==0.13.0
prometheus-client==0.19.0
protobuf==4.25.2
psutil==5.9.8
pyaml==23.12.0
pyarrow==11.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.14
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pymongo==3.11.0
pynndescent==0.5.11
pyod==1.1.3
numpy==1.20.0
pyparsing==3.1.1
pytest==8.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.2
pytz==2023.4
PyYAML==5.4.1
qtconsole==5.5.1
QtPy==2.4.1
querystring-parser==1.2.4
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3fs==2023.12.2
schemdraw==0.15
scikit-base==0.7.2
scikit-learn==1.2.0
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.2
shap==0.44.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==1.4.51
sqlparse==0.4.4
starlette==0.36.3
statsmodels==0.13.5
tabulate==0.9.0
tbats==1.1.3
tenacity==8.2.3
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.3.0
time-machine==2.13.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
tsdownsample==0.1.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.5
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.27.1
waitress==2.1.2
watchdog==4.0.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.1.2
xgboost==1.7.6
xxhash==3.4.1
yarl==1.9.4
yellowbrick==1.5

